+++ week4/LimChaeeun/src/App.css
@@ -1,157 +1,3 @@
-button {
-  background: none;
-  padding: 0;
-  border: none;
-  cursor: pointer;
-}
-* {
-  box-sizing: border-box;
-}
-a {
-  text-decoration: none;
-}
-
-.header-inner {
-  height: 52px;
-  width: 100%;
-  display: flex;
-  justify-content: space-between;
-  align-items: center;
-  padding: 12px 16px;
-  background-color: #fff;
-
-  border-bottom: 1px solid #d9d9d9;
-  position: fixed;
-  top: 0;
-}
-
-.header-inner__title {
-  display: flex;
-  align-items: center;
-  gap: 6px;
-
-  color: #000;
-  font-family: Inter;
-  font-size: 18px;
-  font-style: normal;
-  font-weight: 700;
-  line-height: 28px;
-}
-
-.header-inner__icons {
-  display: flex;
-  align-items: center;
-  gap: 16px;
-}
-
-.content {
-  padding-top: 56px;
-  padding-bottom: 58px;
-}
-
-.content-element {
-  display: flex;
-  padding: 16px 16px 0 16px;
-  flex-direction: column;
-  justify-content: center;
-  align-items: flex-start;
-}
-
-.content-item {
-  display: flex;
-  align-items: stretch;
-  gap: 16px;
-  width: 100%;
-  padding-bottom: 16px;
-
-  border-bottom: 1px solid #eee;
-}
-
-.content-item img {
-  width: 110px;
-  height: 110px;
-}
-
-.content-item__info {
-  width: 100%;
-  display: flex;
-  flex-direction: column;
-  justify-content: space-between;
-  gap: 4px;
-  flex: 1;
-}
-
-.content-item__text {
-  display: flex;
-  flex-direction: column;
-  align-items: flex-start;
-  gap: 4px;
-  flex-grow: 1;
-}
-
-.content-item__title {
-  color: #000;
-
-  font-family: Inter;
-  font-size: 16px;
-  font-style: normal;
-  font-weight: 400;
-  line-height: 20px;
-}
-
-.content-item__meta {
-  color: #8c8c8c;
-
-  font-family: Inter;
-  font-size: 12px;
-  font-style: normal;
-  font-weight: 400;
-  line-height: 16px;
-}
-
-.content-item__price {
-  color: #ff7e36;
-  font-family: Inter;
-  font-size: 15px;
-  font-style: normal;
-  font-weight: 700;
-  line-height: 22px; /* 146.667% */
-}
-
-.content-item__isSold {
-  color: #8c8c8c;
-
-  font-family: Inter;
-  font-size: 15px;
-  font-style: normal;
-  font-weight: 400;
-  line-height: 16px;
-}
-
-.content-item__icons {
-  display: flex;
-  justify-content: flex-end;
-  gap: 4px;
-}
-
-.content-item__icons span {
-  display: flex;
-  align-items: center;
-  gap: 2px;
-  color: #8c8c8c;
-
-  font-family: Inter;
-  font-size: 12px;
-  font-style: normal;
-  font-weight: 400;
-  line-height: 16px;
-}
-
-.content-item__icons img {
-  width: 16px;
-  height: 16px;
-}
-
 .fab {
   padding: 16px;
   position: fixed;
@@ -160,7 +6,6 @@ a {
 }
 
 .fab img {
-  display: flex;
   padding: 12px;
   align-items: flex-start;
   gap: 10px;
@@ -168,36 +13,4 @@ a {
   border-radius: 100px;
   background: #ff7e36;
   box-shadow: 0 4px 12px 0 rgba(0, 0, 0, 0.12);
-}
-
-.bottom-nav {
-  height: 58px;
-  width: 100%;
-  display: flex;
-  padding: 8px 0;
-
-  border-top: 1px solid #d9d9d9;
-  background-color: #fff;
-
-  position: fixed;
-  bottom: 0;
-}
-
-.bottom-nav li {
-  flex: 1;
-}
-
-.nav-item {
-  display: flex;
-  flex-direction: column;
-  align-items: center;
-  justify-content: center;
-  gap: 4px;
-
-  color: #000;
-  font-family: Inter;
-  font-size: 10px;
-  font-style: normal;
-  font-weight: 400;
-  line-height: 14px;
 }
