+++ week4/LimChaeeun/src/components/BottomNav.jsx
@@ -1,0 +1,17 @@
+import NavElement from "./NavElement";
+import navModel from "../model/navModel";
+import "./BottomNav.css";
+
+const BottomNav = () => {
+  return (
+    <nav>
+      <ul className="bottom-nav">
+        {navModel.map((item) => (
+          <NavElement key={item.name} item={item} />
+        ))}
+      </ul>
+    </nav>
+  );
+};
+
+export default BottomNav;
