+++ week4/LimChaeeun/src/components/Content.css
@@ -1,0 +1,107 @@
+.content {
+  padding-top: 56px;
+  padding-bottom: 58px;
+}
+
+.content-element {
+  display: flex;
+  padding: 16px 16px 0 16px;
+  flex-direction: column;
+  justify-content: center;
+  align-items: flex-start;
+}
+
+.content-item {
+  display: flex;
+  align-items: stretch;
+  gap: 16px;
+  width: 100%;
+  padding-bottom: 16px;
+
+  border-bottom: 1px solid #eee;
+}
+
+.content-item img {
+  width: 110px;
+  height: 110px;
+}
+
+.content-item__info {
+  width: 100%;
+  display: flex;
+  flex-direction: column;
+  justify-content: space-between;
+  gap: 4px;
+  flex: 1;
+}
+
+.content-item__text {
+  display: flex;
+  flex-direction: column;
+  align-items: flex-start;
+  gap: 4px;
+  flex-grow: 1;
+}
+
+.content-item__title {
+  color: #000;
+
+  font-family: Inter;
+  font-size: 16px;
+  font-style: normal;
+  font-weight: 400;
+  line-height: 20px;
+}
+
+.content-item__meta {
+  color: #8c8c8c;
+
+  font-family: Inter;
+  font-size: 12px;
+  font-style: normal;
+  font-weight: 400;
+  line-height: 16px;
+}
+
+.content-item__price {
+  color: #ff7e36;
+  font-family: Inter;
+  font-size: 15px;
+  font-style: normal;
+  font-weight: 700;
+  line-height: 22px; /* 146.667% */
+}
+
+.content-item__isSold {
+  color: #8c8c8c;
+
+  font-family: Inter;
+  font-size: 15px;
+  font-style: normal;
+  font-weight: 400;
+  line-height: 16px;
+}
+
+.content-item__icons {
+  display: flex;
+  justify-content: flex-end;
+  gap: 4px;
+}
+
+.content-item__icons span {
+  display: flex;
+  align-items: center;
+  gap: 2px;
+  color: #8c8c8c;
+
+  font-family: Inter;
+  font-size: 12px;
+  font-style: normal;
+  font-weight: 400;
+  line-height: 16px;
+}
+
+.content-item__icons img {
+  width: 16px;
+  height: 16px;
+}
