+++ week4/LimChaeeun/src/components/Header.css
@@ -1,0 +1,32 @@
+.header-inner {
+  height: 52px;
+  width: 100%;
+  display: flex;
+  justify-content: space-between;
+  align-items: center;
+  padding: 12px 16px;
+  background-color: #fff;
+
+  border-bottom: 1px solid #d9d9d9;
+  position: fixed;
+  top: 0;
+}
+
+.header-inner__title {
+  display: flex;
+  align-items: center;
+  gap: 6px;
+
+  color: #000;
+  font-family: Inter;
+  font-size: 18px;
+  font-style: normal;
+  font-weight: 700;
+  line-height: 28px;
+}
+
+.header-inner__icons {
+  display: flex;
+  align-items: center;
+  gap: 16px;
+}
