+++ week4/LimChaeeun/src/reset.css
@@ -122,3 +122,15 @@ table {
   border-collapse: collapse;
   border-spacing: 0;
 }
+button {
+  background: none;
+  padding: 0;
+  border: none;
+  cursor: pointer;
+}
+* {
+  box-sizing: border-box;
+}
+a {
+  text-decoration: none;
+}
